[DB] Add open check for debug builds, qint64 for count

+++ headers/utils/db_connector.h
@@ -75,7 +75,7 @@ public:
     bool deleteTable(const std::string &name);
     bool tableExists(const std::string &table);
     bool dropTable(const std::string &table);
-    int count(const std::string &table, const std::string &where = "");
+    qint64 count(const std::string &table, const std::string &where = "");
     std::string file() const;
     bool isOpen() const;
     std::vector<std::string> tableNames();
